requirements.txt
streamlit==1.35.0
tensorflow==2.16.1
pillow==9.5.0

